use static::define() instead of self::$container
when iterating over self::$container, it was an empty array,
which lead to no events returned in getSubscribedEvents()

As the events are defined in define(), we can simply use the array
returned from there.

To do this, define() needs to be static, too.

As some PHP Versions disallow abstract static methods,
change it to a non-static version and add a sane default
implementation.